fix(alerts): fix PostgreSQL NUMERIC to Rust f64 conversion in alert evaluator
- Add ::float8 cast to AVG() and percentile_cont() queries for sqlx compatibility
- PostgreSQL AVG returns NUMERIC type which sqlx couldn't decode to f64
- Also includes: ServerSetupModal English translation, VITE_API_URL fix
- Add test-alerts.sh script with parallel calls for faster alert testing

+++ backend/src/jobs/alert_evaluator.rs
@@ -145,7 +145,7 @@ async fn get_metric_value(pool: &PgPool, rule: &AlertRuleEval) -> Result<f64, sq
         ),
         "avg_latency" => format!(
             r#"
-            SELECT COALESCE(AVG(latency_ms), 0.0) as value
+            SELECT COALESCE(AVG(latency_ms)::float8, 0.0) as value
             FROM request_metrics 
             WHERE time > NOW() - INTERVAL '{}'
             {}
@@ -155,7 +155,7 @@ async fn get_metric_value(pool: &PgPool, rule: &AlertRuleEval) -> Result<f64, sq
         "p95_latency" => format!(
             r#"
             SELECT COALESCE(
-                percentile_cont(0.95) WITHIN GROUP (ORDER BY latency_ms),
+                percentile_cont(0.95) WITHIN GROUP (ORDER BY latency_ms)::float8,
                 0.0
             ) as value
             FROM request_metrics 
@@ -177,7 +177,6 @@ async fn get_metric_value(pool: &PgPool, rule: &AlertRuleEval) -> Result<f64, sq
     };
 
     let row = sqlx::query(&query).fetch_one(pool).await?;
-
     let value: f64 = row.try_get("value").unwrap_or(0.0);
     Ok(value)
 }
